requirements.txt
tensorflow==2.15.0
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
matplotlib==3.8.2
seaborn==0.13.0
Pillow==10.1.0
